feature: Support for building result with setattr
_more: So far, the converter would always generate a list of values and
then pass them to the constructor of the target type. From now on, there’s
an additional `strategy` argument in `mappr.convert`. It defaults to
`mappr.Strategy.CONSTRUCTOR` which is the same as it’s been so far.
What’s new is the ability to set it to `mappr.Strategy.SETATTR` which will
cause the converter to use empty constructor and then use `setattr()` to
set all the values on newly created object. Might cater to some cases.