feature: Can specify default strategy for the mapping
_more: Until now, you could only specify the strategy during conversion,
so if you wanted to always use a SETATTR strategy you would have to
pass it each time you convert something. Now you can define the strategy
when registering a mapping and then it will be used as default for all
conversions. You can still override the strategy directly in `convert`,
but the default is now specified in the mapping.